chore(specs): fields features vervangen met expressieve scenarios
