+++ features/step_definitions/gegeven-stepdefs-gezagsrelaties.js
@@ -1,5 +1,7 @@
 const { Given } = require('@cucumber/cucumber');
 const { createObjectFrom } = require('./dataTable2Object');
+const { getBsn, getPersoonMetAanduiding } = require('./contextHelpers');
+const { createGezag } = require('./gezagsrelatie');
 
 Given(/^voor de persoon met burgerservicenummer '(\d*)' gelden de volgende gezagsrelaties$/, function (burgerservicenummer, dataTable) {
     if(this.context.gezag === undefined) {
@@ -62,4 +64,69 @@ Given(/^het gezag heeft geen derden$/, function() {
 
     let gezag = persoon.gezag.at(-1);
     gezag.derden = [];
+});
+
+function gegevenDePersoonHeeftGezagsrelaties(aanduiding) {
+    if(this.context.gezag === undefined) {
+        this.context.gezag = [];
+    }
+    this.context.bsnCurrentGezagPersoon = getBsn(getPersoonMetAanduiding(this.context, aanduiding));
+
+    if(!this.context.gezag.find(g => g.burgerservicenummer === this.context.bsnCurrentGezagPersoon)) {
+        this.context.gezag.push({
+            burgerservicenummer: this.context.bsnCurrentGezagPersoon,
+            gezag: []
+        });
+    }
+}
+
+Given('{aanduiding} heeft de volgende gezagsrelaties', gegevenDePersoonHeeftGezagsrelaties);
+Given('{aanduiding} heeft geen gezaghouder', gegevenDePersoonHeeftGezagsrelaties);
+
+Given('eenhoofdig ouderlijk gezag over {aanduiding} met ouder {aanduiding}', function (minderjarige, ouder) {
+   let persoon = this.context.gezag.find(g => g.burgerservicenummer === this.context.bsnCurrentGezagPersoon);
+
+   persoon.gezag.push(createGezag(this.context, 'eenhoofdig ouderlijk gezag', minderjarige, ouder, undefined));
+});
+
+Given('gezamenlijk ouderlijk gezag over {aanduiding} met ouder {aanduiding} en ouder {aanduiding}', function (minderjarige, ouder1, ouder2) {
+   let persoon = this.context.gezag.find(g => g.burgerservicenummer === this.context.bsnCurrentGezagPersoon);
+
+   persoon.gezag.push(createGezag(this.context, 'gezamenlijk ouderlijk gezag', minderjarige, ouder1, ouder2));
+});
+
+Given('gezamenlijk gezag over {aanduiding} met ouder {aanduiding} en derde {aanduiding}', function (minderjarige, ouder1, ouder2) {
+   let persoon = this.context.gezag.find(g => g.burgerservicenummer === this.context.bsnCurrentGezagPersoon);
+
+   persoon.gezag.push(createGezag(this.context, 'gezamenlijk gezag', minderjarige, ouder1, ouder2));
+});
+
+Given('gezamenlijk gezag over {aanduiding} met ouder {aanduiding} en een onbekende derde', function (minderjarige, ouder1) {
+   let persoon = this.context.gezag.find(g => g.burgerservicenummer === this.context.bsnCurrentGezagPersoon);
+
+   persoon.gezag.push(createGezag(this.context, 'gezamenlijk gezag', minderjarige, ouder1, undefined));
+});
+
+Given('voogdij over {aanduiding}', function (minderjarige) {
+   let persoon = this.context.gezag.find(g => g.burgerservicenummer === this.context.bsnCurrentGezagPersoon);
+
+   persoon.gezag.push(createGezag(this.context, 'voogdij', minderjarige, undefined, undefined));
+});
+
+Given('voogdij over {aanduiding} met derde {aanduiding}', function (minderjarige, ouder1) {
+   let persoon = this.context.gezag.find(g => g.burgerservicenummer === this.context.bsnCurrentGezagPersoon);
+
+   persoon.gezag.push(createGezag(this.context, 'voogdij', minderjarige, ouder1, undefined));
+});
+
+Given('tijdelijk geen gezag over {aanduiding} met de toelichting {toelichting}', function (minderjarige, toelichting) {
+   let persoon = this.context.gezag.find(g => g.burgerservicenummer === this.context.bsnCurrentGezagPersoon);
+
+   persoon.gezag.push(createGezag(this.context, 'tijdelijk geen gezag', minderjarige, undefined, undefined, toelichting));
+});
+
+Given('gezag over {aanduiding} is niet te bepalen met de toelichting {toelichting}', function (minderjarige, toelichting) {
+   let persoon = this.context.gezag.find(g => g.burgerservicenummer === this.context.bsnCurrentGezagPersoon);
+
+   persoon.gezag.push(createGezag(this.context, 'niet te bepalen', minderjarige, undefined, undefined, toelichting));
 });
