+++ scripts/specs-verify.sh
@@ -63,5 +63,19 @@ npx cucumber-js -f json:./test-reports/cucumber-js/personen/test-result.json \
                 features/zoek-met-nummeraanduiding-identificatie \
                 features/zoek-met-postcode-en-huisnummer \
                 features/zoek-met-straatnaam-huisnummer-en-gemeente-van-inschrijving \
-                --tags "not @skip-verify" \
-                --world-parameters "$PARAMS"
+                -p DataApi
+
+npx cucumber-js -f json:./test-reports/cucumber-js/personen/test-result.json \
+                -f summary:./test-reports/cucumber-js/personen/test-result-summary.txt \
+                -f summary \
+                features/gezag-persoon-beperkt \
+                features/persoon \
+                features/persoon-beperkt \
+                features/raadpleeg-met-burgerservicenummer \
+                features/zoek-met-adresseerbaar-object-identificatie \
+                features/zoek-met-geslachtsnaam-en-geboortedatum \
+                features/zoek-met-geslachtsnaam-voornamen-en-gemeente-van-inschrijving \
+                features/zoek-met-nummeraanduiding-identificatie \
+                features/zoek-met-postcode-en-huisnummer \
+                features/zoek-met-straatnaam-huisnummer-en-gemeente-van-inschrijving \
+                -p DataApiDeprecated
